Changed getStandardRate in Currency to return a float rather than a int
Also solved every conflict it generated

+++ src/payStation/impl/PayStationImpl.java
@@ -54,7 +54,7 @@ public class PayStationImpl implements PayStation {
 
 	@Override
 	public int getMinutes() {
-		int secondsStandardRate = payment * currency.getStandardRate();
+		int secondsStandardRate = (int)(payment * currency.getStandardRate());
 		return payRate.calculateTime(secondsStandardRate);
 	}
